Refactor: Streamline BlobController by moving blob operations to BlobService
